Set `aria-hidden` to true when popup is closed
Currently when a user closes the popup, it disappears from the screen
but is still picked up by the screen reader. This is because all we do
is change the popup's opacity and essentially make it invisible when
closed. This PR sets `aria-hidden` to true so we don't confuse screen
reader users. The reason why we use `aria-hidden` over `display: none`
is because the popup is animated and it would be a lot more complicated
to use the later.

DAC ticket: https://trello.com/c/aQqSISwt/160-bleedthrough-issue-id-dacbleedthrough01 🐿 v2.12.5